Rework createTree
